refactor: MatchResult card 패키지로 이동
