+++ src/main/java/domain/card/MatchResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum MatchResult {
